requirements.txt
absl-py==2.0.0
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.1.0
autopep8==2.0.4
cachetools==5.3.2
ccxt==4.1.79
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.1
gast==0.5.4
google-auth==2.25.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
gym==0.26.2
gym-notices==0.0.8
h5py==3.10.0
idna==3.6
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
pandas-ta==0.3.14b0
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.23.4
psycopg2==2.9.9
pyaml==23.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycares==4.4.0
pycodestyle==2.11.1
pycparser==2.21
pyparsing==3.1.1
pytab==1.0.4
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scikit-optimize==0.9.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.23.0
termcolor==2.4.0
threadpoolctl==3.2.0
torch==2.1.1
ttkbootstrap==1.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vaderSentiment==3.3.2
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
xgboost==2.0.2
yarl==1.9.4
